tree-wide: prune most files not touched by bbr_stan_alpha branch
Delete most of the files that aren't touched by the bbr_stan_alpha
branch.  Some untouched files that we know we'll want (e.g.,
.Rbuildignore, R/utils-pipe.R) are retained.

This could also be done after the merge of bbr_stan_alpha.  The
advantage of doing it before is that, if we accidentally deleted a
file that was touched by the branch, it will be flagged by a merge
conflict.