tree-wide: prune unchanged and bbr-side bits
An earlier commit pruned files that were _unmodified_ between the
bbr_stan_alpha branch and main.  However, within the modified files,
there are lots of parts that are either unchanged or things that will
need to be on the bbr-side (at least until there is a wider rework).

Prune these within-file bits now.

Note that this introduces breakage that will need to be resolved by
both changes in this repo and changes in the bbr repo.