001-bbr.R: import private bbr things
There are a handful of private functions and constants that this
package needs from bbr.  At least for the functions, it'd be good to
keep this set small, and we may want to look into reducing the current
set by deciding that it's fine to export some of these and reworking
others to build in a model-type-specific dispatch.

Note that, on the bbr side, all of these are marked with a "SHARED
..." comment.