open.R: add file.edit wrapper to silence R CMD check
Check will complain with a plain file.edit():

  Undefined global functions or variables:
    file.edit

However, using utils::file.edit() isn't a good option because it
bypasses RStudio's file.edit() override, leading to different (and in
this context undesirable) behavior: the file is opened as a popup that
the user must save before continuing.

Add a kludge to avoid the warning.