run_chains: suppress Rd generation
The plan is for run_chains() to be an unexported helper for
submit_model.bbi_nmbayes_model().

Note that using noRd rather than 'keywords internal' follows the
style.tidyverse.org and rOpenSci guidelines, but it deviates from what
bbr does.  If there ends up being a good reason to generate the Rd,
it's of course easy enough to switch.