drone: hide stdout for cmdstan installation unless it fails
Building cmdstan installation floods the output with over ~150 lines,
which makes it hard to find the output that the main 'R CMD check'
that people are most often looking for.

Redirect standard output to a file and only show it if installation
fails.  (Another option would be to do the installation in a separate
step, but then we'd need to install cmdstan to a spot that persists
between steps.)

Note cmdstanr emits some higher-level message on stderr, so there's
still some indication of what's happening.