bayes/1100.ctl: make more suitable for testing
This control stream was copied from the NONMEM Bayes example project.
The plan is to run it and track its output within inst/ so that we can
test things like extracting draws without needing to run the model
first.  However, at the moment it runs too slowly for testing
purposes (over an hour for running one chain locally) and the
resulting output is too large, in particular the *.iph files (~25 M
per chain).

To reduce the time and resulting size:

 * decrease the number of iterations

 * cut the number of chains to two

 * switch from NUTS to BAYES

That above is bad in terms of getting good samples/estimates, but
that's not relevant for the tests.  With these changes, the run takes
under a minute, and the iph files are ~2 M.