tests: tweak descriptions in preparation for nmbayes
A lot of the test descriptions follow the form
"methodx.bbi_stan_model() ...".  That's a bit awkward once nmbayes is
introduced because many of those just rely on the parent method to
work (i.e. there's no methodx.bbi_nmbayes_model()).  Plus, the reader
has to scan to the middle of the description to spot the difference
between otherwise identical stan and nmbayes test descriptions.

For tests that are going to have nmbayes analogs (and a few others),
rewrite the descriptions to use a "stan:" prefix.  While touching
these lines, add parentheses to function names for consistency with
other descriptions.