nmbayes-draws: optionally include iph data in draws
NONMEM writes samples of phi/eta to *.iph when the estimation
specifies BAYES_PHI_STORE=1.  Unlike the ext files, the format needs
to be reshaped a bit before converting it to a draws object because
there are multiple rows per iteration.

Do the reshaping by using dplyr/tidyr to construct a data frame where
the ID and SUBPOP values are encoded as indices in the parameter
names.  Some notes:

 * The combination of SUBPOP along with either ID or SUBJECT_NO should
   uniquely identify rows for an iteration.  The difference between ID
   and SUBJECT_NO is that ID is the user-specified value while
   SUBJECT_NO is NONMEM's internal index/label.

 * The 1-based integer index of ID and SUBPOP are used instead of the
   label itself.  The main advantage of doing this is that, if we
   construct the names as "par[M,N,...]", posterior can handle it via
   its rvars functionality.

   The downside is that the values, in particular the ID, don't have
   their user-specified value.  To make it easier to convert the index
   back to the original value, an upcoming commit will store tibbles
   with the mappings under a custom draws object's attribute.

 * For non-mixture models, SUBPOP is always 0.  In that case, we could
   drop the index.  However, keeping it means that the caller gets
   more predictable output.  Once users have extracted rvars from
   these values, they can drop the redundant dimension with
   posterior's drop() method if desired.

 * The current approach may be too slow for larger data sets.  It will
   be sped up a bit in upcoming commits, but, at the very least, it
   provides a reference and tests that a rewrite can be developed
   against.

Include the iph data in the same draws object as the ext data because
that's consistent with Stan's output/approach (although there's no
reason we couldn't keep them separate, if most users end up preferring
that).

Closes #34.