nmbayes-draws: store ID/subpop maps as draws object attribute
As described a couple of commits back, the ID and SUBPOP values are
encoded as 1-based indices in the parameter names.  Store tibbles
mapping these indices to the original values under an attribute in the
draws object.  (Use "nmbayes" as the attribute in anticipation of
storing other metadata and to hopefully reduce the change of conflicts
with posterior's attributes down the road.)

A custom class is needed so that the attribute is consistently
included when posterior::as_draws_* functions are called directly on a
model object.  Note that it'd be nice to reduce the remaining
duplication in these methods, but I didn't see an obvious way to do so
given that 1) the object needs to be modified before the NextMethod()
call and 2) the NextMethod() call must happen within the method.