nmbayes_draws: restrict iph draws to ETAs and MCMCOBJ
Reading the draws from the Bayes example project (model 1100) with
read_fit_model() takes a little over 8 seconds.  Based on the example
project, it seems like ETAs and MCMCOBJ are often the only values of
interest, so restrict the initial fread to those.  This brings the
read_fit_model() call to under 5 seconds.

The plan is to eventually extend nmbayes_draws() to select a subset of
variables.  When that's done, we should consider whether to also
expose an option to select which iph variables are selected.