reshape_iph: reposition mutate for a speedup
reshape_iph() makes a long table from the raw iph data and then
creates ID and SUBPOP index names.  However, this operation can be
done before making the table long, reducing the number of values that
as.integer(factor(...)) needs to consume.

This saves about half a second when testing on the Bayes example
project (model 1100).