shrinkage tests: check shape of results
shrinkage.rvar() accepts errors with arbitrary dimensions, calculating
shrinkage over the specified group margins.  Add tests to make sure
that the expected shape comes back.

This of course doesn't say anything about whether the value is
correct.  At the moment, the only test that provides support for that
is the comparison of the shrinkage() values to the values of the
baysh() the reference implementation.