shrinkage: make caller-supplied variance optional
The example project's baysh() helper gets the variance of the errors
from diagonal of the OMEGA matrix.  However, the referenced Gelman and
Pardoe paper calculates it from the errors directly.  For
shrinkage.draws() and shrinkage.rvar(), do the same if the caller
doesn't specify the value.

Don't bother exposing this via shrinkage.bbi_nmbayes_model() until
it's clear there's a case where using the errors directly would be
more appropriate than the using OMEGA-based values.  The OMEGA-based
values are likely more reliable, particularly in the context of a
small number of groups.