drone: set up cmdstan for release phase
The release phase is failing to generate the vignettes:

  Caused by error:
  ! CmdStan path has not been set yet. See ?set_cmdstan_path.

https://github-drone.metrumrg.com/metrumresearchgroup/bbr.bayes/125/2/3

Unlike the main test step, the build step of the release phase doesn't
install cmdstan.  I'm not sure why the above error is triggered now
but wasn't for the last dev tag (0.0.0.9006).

Try to resolve this error by copying the cmdstan setup from the main
test phase.