shrinkage tests: compare more .rvar results to ref implementation
For shrinkage.bbi_nmbayes_model(), the result can be compared directly
to the example project reference implementation to give some
confidence that the correct values are being calculated.  That NONMEM
model sends in 2d input and triggers the multidimensional code path
using the last dimension as the group_idx, but we don't have a
one-to-one reference for the 1d code path or for multidimensional
values with a non-default group_idx.

However, we can compare these cases to the reference result by
selecting particular input and doing more targeted assertions.

 * To trigger the 1d code path in a way that can be compared, call
   shrinkage.rvar() with rvars constructed for sub-elements of ETA and
   OMEGA, and then compare that to each scalar in the ref result.

 * Check a simple 3d case by feeding the ETA values to
   shrinkage.rvar() keeping the redundant dimension (unlike
   shrinkage.bbi_nmbayes_model())

 * To test the calculation over a non-default group_idx, duplicate the
   values along an additional dimension and check that both the
   sub-elements match the reference result.