shrinkage.bbi_nmbayes_model: don't relay dots
shrinkage.bbi_nmbayes_model() passes `...` on to shrinkage.rvar(), but
there's not a case where we would want the caller to do this.
Specifying any other group_idx would be invalid.

Drop the dots, and add a check_dots_used() call so that callers get an
informative error if they try to pass group_idx.