stan: add function to copy stan_gq model from regular model
A couple of commits back added a copy_model_from() method for
bbi_stan_gq_model objects.  However, that's for copying from the
_same_ model type, and it will often be useful to copy stan_gq models
from regular stan models.

Add a dedicated function for copying a stan model to a stan_gq model.
This involves some special handling:

 * adjust the set of copied files (we do not want to copy -init.R)

 * drop most values from the parent's -stanargs.R, keeping the seed
   for convenience

 * tweak the type of the parent model so that copy_model_from_impl()
   returns stan_gq as the new model's type

   This is necessary because copy_model_from_impl() takes the type
   directly from the parent class.  (It might be worth extending
   copy_model_from_impl() to let the caller specify the desired type.)

We may want to hook into this spot later to add more special
handling (e.g., adding a dedicated tag or linking data instead of
copying it).

Rather than add a helper that does the copy, we could instead take the
approach of nmbayes_convert() and leave the copy up to the caller:

  copy_model_from(...) %>% *_convert()

That's nice from a composability standpoint, but it means that we need
to revert and adjust files once they're on disk rather than taking
advantage of copy_model_from_impl's setup_fn to get directly to the
target state.  Also, because we're not the ones doing the copy, we
need to worry about whether the model is in a valid state (e.g., has
the caller already tweaked the stanargs.R or did they accidentally
pass an already executed model that has outputs?).

(We should consider replacing nmbayes_convert() with
copy_nonmem_model_as_nmbayes() or similar for consistency and because
to some degree the same concerns above apply to nmbayes models.)