tests: add local_tempdir() wrapper that returns normalized paths
For better or worse, bbr is fairly aggressive with normalizing paths.
It's not uncommon to have a TMPDIR value that won't match the result
of normalizePath() (e.g., if it contains a symlink), tripping up tests
comparisons and also bbr logic (e.g., get_based_on()).

As an example, here is how the copy_model_from() call in
test-build-data.R fails on a system where TMPDIR is under /var and
/var is a symlink to /private/var:

  Error: Parsed 1 models as `based_on` but cannot find .yaml files for 1
  of them:
  /private/var/{rest of temp path}/../../../../../../{full path to repo}/inst/model/stan/fxa
  Backtrace:
   1. bbr::copy_model_from(STAN_MOD1, tempfile())
        at test-build-data.R:30:2
   2. bbr.bayes:::copy_model_from.bbi_stan_model(STAN_MOD1, tempfile())
        at bbr/R/copy-model-from.R:73:2
   3. bbr:::copy_model_from_impl(...)
        at bbr.bayes/R/copy-model-from.R:19:2
   4. bbr::new_model(...)
        at bbr/R/copy-model-from.R:182:2
   5. bbr::replace_all_based_on(.mod, .based_on)
        at bbr/R/new-model.R:74:2
   8. bbr:::safe_based_on(get_model_working_directory(.mod), .based_on)
        at bbr/R/modify-model-field.R:298:2
   9. bbr:::strict_mode_error(...)
        at bbr/R/modify-model-field.R:493:4

Sidestep these issues (which aren't for bbr.bayes to solve) by adding
a local_tempdir() wrapper that returns a normalized path.