stan: move scaffolds to standalone files
The scaffolds used by the add_*_file() functions are stored as
strings, which (at least for me) makes it harder to view and tweak
these (due to the lack of things like automatic indentation and
keyword highlighting).  Move these scaffolds to dedicated files in
inst/ so that they can be handled as regular .R and .stan files.