add-file-to-model-dir.R: support stan_gq models
At the directory level, the main difference between regular
bbi_stan_model objects (i.e. models destined for $sample()) and
bbi_stan_gq_model objects (i.e. models destined for
generate_quantities()) is that the former has -init.R files and the
latter -fitted-params.R.

Define an add_stan_*_file() helper for -fitted-params.R.

With that addition, new_model(..., .model_type = "stan_gq") is now
functional.  Down the road, it may make sense to add a .stan scaffold
file that is more appropriate for defining a standalone generated
quantities model.  On the other hand, new_model() is probably not
going to be a common way to add a stan_gq model (versus the
copy_stan_model_as_gq() function that will be added later), so it may
not be worth the trouble.

Note that the add_file_to_model_dir() docs point to the bbi_stan_model
docs.  Those will be updated for stan_gq models in a later commit.