check_up_to_date: define method for stan_gq models
A tailored check_up_to_date() method is needed for stan_gq models
because -fitted-params.R needs to be considered and -init.R should not
be considered.

The new check only inspects the hash of the -fitted-params.R file, so
it won't report the model as stale when the upstream fit has changed.
Given that -fitted-params.R may grab a fit from anywhere and return
anything accepted by generate_quantities(), doing a more extensive
check that handles all cases is tricky, but perhaps we'll want to
extend this to handle the common case of returning a vector of
CSVs (e.g., we could hash the hashes of each file).  That could be
expensive, so should be optional (like .build_data).

Note that we don't use bbi_stan_*summary objects yet, so there's not
much point in adding a method for bbi_stan_gq_summary, but do so for
consistency with bbi_stan_model objects.

Corresponding tests will be added once an upcoming commits adds a
stan_gq model under inst/model/.