model_diff: support stan_gq models
The model_diff.bbi_stan_model() method tweaks the allowed values of
the .file argument to support stan models.  For stan_gq models,
-init.R is no longer relevant and -fitted-params.R is.

Adding a dedicated method for bbi_stan_gq_model objects doesn't buy us
much because 1) it's meaningful to diff between stan and stan_gq
models (e.g., to see how the data file change) and 2) using a method
makes it seem like the type of the first argument is more important
than the type of the second in determining what can be diffed, which
isn't the case.

So, add handling for stan_gq models within the bbi_stan_model method.
Note diffobj::diffFile() does not gracefully deal with non-existent files
(e.g., by showing one side as a new or deleted file), so we can't just
send init or fitted-params files through to the diff call if one model
is stan and the other is stan_gq.