copy_model_from.bbi_stan_model: pull logic into helper
The upcoming stan_gq method will need the same logic but with a
different set of files to copy to the new model.  Add a helper that
can be used by the stan_gq method too.

Repeating the large set of arguments again is ugly, but I don't see a
good way to avoid that.