add_file_to_model_dir_impl: drop inaccessible .overwrite argument
The internal add_file_to_model_dir_impl() has an .overwrite argument
to control whether it will clobber an existing destination path.  If
the destination file exists and the default .overwrite is
FALSE (default determined by .source_file), an add_*_file() caller
will see a message sayint to "pass `.overwrite = TRUE`".  However,
none of the exported functions that use add_file_to_model_dir_impl()
expose this argument, so the message is inappropriate and confusing.

Given it's not used, drop the .overwrite argument.  Note that, if we
decide to add it back and expose it, the .source_file default should
be rewritten to not use a vectorized ifelse (it could instead be
'!is.null(.source_file)').