submit_model.bbi_stan_model: drop unsupported args
submit_model.bbi_stan_model() has four parameters that it ignores
after warning if they are not NULL.  This means that they are
advertised in the submit_model.bbi_stan_model() documentation without
any mention that they are invalid, which is confusing.

Instead of giving warning a error, drop these parameters from the
method.  If caller provides one, they'll now see an error like this:

  > submit_model(m, .bbi_args = list("foo"))
  Error in `submit_model()`:
  ! `...` must be empty.
  ✖ Problematic argument:
  • .bbi_args = list("foo")
  Run `rlang::last_error()` to see where the error occurred.