shrinkage: choose more appropriate dots check
All of the shrinkage methods call rlang::check_dots_used().  Although
the end result is okay (giving an error if a user tries to pass an
argument via ...), check_dots_used() is doing a more involved check
than needed here (these methods do not relay the dots) and leads to
unnecessary work (the error happens when the method exits).

Switch to using check_dots_empty() to do an upfront check.