stan: add helpers for getting/setting gq_parent field
There isn't an explicit link between a stan_gq model and the plain
stan model that it was (usually) copied from.  In many cases, the
current taking the leftmost get_based_on() value will point to the
upstream stan model.  However, we can't rely on that to give the right
answer (e.g., if you copy a stan_gq model from another stan_gq, you
need to call `get_based_on(get_based_on(m)[1])[1]`).

In the case of -fitted-params.R, that's acceptable because the code
there is meant as a useful starting point that the user can adjust the
code as needed.  But the plan is to make submit_model() record the
hash of the upstream stan model's bbi_config.json in the stan_gq
bbi_config.json for check_up_to_date().  In that case, it's more
important to have a value whose explicit purpose is to track the link
to the stan model whose results are used as the fitted params.

Add helpers for getting and setting the gq_parent field.  Note that
get_stan_gq_parent() is really the only one that we _need_ to expose
(for use in -fitted-params.R).  copy_model_as_stan_gq() and
copy_model_from.bbi_stan_gq_model() will learn how to set this field,
and it won't need to be adjusted or set manually in most cases.
However, if we're going to add a field that is central to the model
definition, it seems worth adding a set of helpers that mirror bbr's
existing helpers (in particular get_based_on(), add_based_on(),
replace_all_based_on(), and remove_based_on()).

Note that the documentation for these will be extended in an upcoming
commit.