copy-model-from.R: set gq_parent field
As mentioned in the previous commit, the gq_parent field is going to
be used as an explicit link between a stan_gq model and the parent
stan model.  Teach copy_model_as_stan_gq() to point this field to the
parent stan model, and teach copy_model_from.bbi_stan_gq_model() to
propagate the gq_parent field from the parent stan_gq model.