submit_model.bbi_stan_gq_model: flag missing gq_parent config
The gq_parent field points to a regular Stan model (or models) whose
fit is being fed into the standalone generated quantities run.  So
that model needs to have already been executed and, for
check_up_to_date() purposes, we rely on being able to hash the
resulting bbi_config.json.  Abort early if something looks off.

Note that, for long running models, there's an extended window where
the gq_parent could be changed after this guard and before the
get_gq_parent_md5() following the model run.  At some point, it may be
worth reordering things to collect hashes earlier so that the model
would at least be reported out of date.  However, that's 1) not
specific to the gq_parent hash and 2) would involve deliberate action
on the user's part and may be outside of the sort of mistakes
bbr/bbr.bayes should guard against.