check_stan_model: also check .stan syntax by default
check_stan_model() looks in the model directory for the required files
and checks that stub scaffolds have been updated.  However, given the
function's name, it's likely that users will expect it to flag issues
in the .stan model.  That's cheap to do via cmdstanr's $check_syntax()
method, which calls stanc on the .stan file, so update it to do so by
default.

Note that the advantage here is mostly just giving direct
check_stan_model() callers more information; the stanc call would be
the first step of the `make` call triggered by $compile(), so a user
calling check_stan_model() upfront won't save much time over calling
submit_model() directly.

This commit holds off on a couple of things:

 * it does not expose $check_syntax() arguments.  There may be cases
   where users want pedantic, stanc_options, or include_paths.  They
   can still access these by grabbing the model object with

     cmdstanr::cmdstan_model(get_model_path(.mod), compile = FALSE)

   and calling $check_syntax() directly.

   For the latter two, the bigger issue is how bbr.bayes should
   support specifying these and other options when calling $compile()
   in submit_model().

 * it does not add a check_stan_model() argument for triggering
   $compile().  It's unclear to me how often a user would want to do
   this ahead of submit_model().  If check_syntax() passes, are there
   regular problems that are triggered by $compile()?  If not common,
   having the user grab the CmdStanModel object with the snippet above
   seems like an acceptable approach.

   And if calling $compile() separately from submit_model() is a
   common need, it probably makes more sense to add a dedicated
   function (e.g., exporting compile_stanmod) rather than extending
   check_stan_model() with something that's more than a check.