read_fit_model.bbi_stan_model: signal custom error if RDS is missing
readRDS() emits a warning and errors if the RDS doesn't exist:

  Error in gzfile(file, "rb") : cannot open the connection
  In addition: Warning message:
  In gzfile(file, "rb") :
    cannot open compressed file 'fit.RDS', probable reason 'No such file or directory'

The next commit will read_fit_model() alls where it's okay if a fit
doesn't exist.  Catch whether fit.RDS doesn't exist upstream so that
we don't have to worry about handling an error and a warning, and use
a custom error class so that we can reliably detect this case without
matching messages or indiscriminately swallowing errors.