treewide: don't qualify tidyselect helpers
Various spots use tidyselect:: for selection helpers like
starts_with().  Doing that is confusing because these are actually
part of the tidyselect DSL and not interpreted by R.  Also, some spots
qualify all_of() with tidyr, which is more confusing because tidyr
just re-exports that from tidyselect.

Drop the prefix, and import all of the used helpers in aaa.R to
silence 'R CMD check'.