import_stan_init: set seed when calling make_init
If make_init() relies on the random seed to generate a list of lists
as the return value, the subsequent sampling doesn't generate the same
values unless the user sets the seed within the make_init() call.

Make this case reproducible without user intervention by setting the
seed around the make_init() call.  We have a seed value for a sample
call, so just use that here too.