import_stan_init: call function returned by make_init
cmdstanr's $sample() will accept a function for its init argument.
However, it's up to the user to set the seed to make the sampling
reproducible.  For bbr.bayes, we want users to be able to get the same
result by only setting the mandatory stanargs seed [*].

Make this the case by calling the returned function ourselves, with
the seed set to the standargs value, rather than leaving it to
cmdstanr's process_init_function().

Another option would be to create a wrapper function around the
returned function that sets the seed, passing that wrapper to
cmdstanr.  That has the appeal of passing the same type of value to
$sample() and minimizing the logic that we're duplicating from
cmdstanr.  However, we plan to add a helper for previewing the init
values, so we're going to need to duplicate that logic anyway.

[*] ... assuming the set of factors described at
    https://mc-stan.org/docs/reference-manual/reproducibility.html are
    held constant

Closes #19