check_stan_model: suppress 'incomplete final line' warning
To check syntax, check_stan_model() calls cmdstanr::cmdstan_model().
That leads to cmdstanr calling readLines(), which will give a
"incomplete final line" warning if the file does not have a trailing
newline.

Although ideally the user would put a newline at the end of the file,
it's valid syntax to leave it off (and RStudio makes it easy to do),
so hide the warning to avoid check_stan_model() callers mistaking it
for a complaint coming from check_stan_model().