stan_summary_log: fix non-scalar variable selection
Passing a value like "beta" to the variables argument of cmdstanr's
$summary() can select a scalar variable ("beta") but also non-scalar
expansions (e.g., "beta[1]", "beta[2]", ...).  The intent was for the
variable argument of stan_summary_log() to behave the same, but the
non-scalar selection doesn't work because extract_param_summaries()
intersects the specified names with the fit's
$metadata()[["variables"]], which has the expanded names.

Add $metadata()[["stan_variables"]] to the set to enable non-scalar
selection.

Closes #63.