safe-source: restrict evaluation environment
To create the environment for source(), safe_source_function() calls
new.env() with the default value for `parent`, leading to the parent
of the environment being the execution environment of
safe_source_function().  That means that the evaluated function (e.g.,
make_standata) has access to any bindings in namespace:bbr.bayes and
its parents (most notably imports:bbr.bayes and the global
environment).

The sourced file should not have access to bbr.bayes namespace.
Walking up the chain, some options for a more appropriate environment
are

 1) the global environment

 2) the parent of the global environment

 3) the environment for the first base package encountered (e.g.,
    package:stats)

 4) the base environment

*1* should be intuitive for the caller because it follows what they
would see if they called make_standata() and friends directly in the
session.  The downside is that the sourced code can access bindings
from the global environment, opening up the possibility for bugs.

Using the base environment (*4*) would prevent this, but it'd also
prevent the user from accessing the package environments on the search
path, including the base packages.  They'd have to qualify all package
references (e.g., "stats::rnorm").

*2* keeps everything except for the global environment on the search
path.  That eliminates the most likely source of leaks, but the
evaluation can still be influenced by non-standard things that the
user has attached (most likely packages, but this could also be data
frames or non-package environments).

*3* could work okay most of the time, forcing the caller to qualify
names for anything other than base packages.  However, we can't assume
the set of base packages is the same across users (it can be
configured by the defaultPackages option and R_DEFAULT_PACKAGES) or
that a base package comes upstream of non-base packages in the search
path (e.g., it won't if a non-base package is loaded in .Rprofile).

Go with *2* as a simple compromise between *1* and *4* that should
eliminate the most common source of leaked outside state.

Closes #25.