vignettes/getting-started-stan.Rmd: consistently override bbr links
For package A to auto-link to package B's docs, package B needs to put
its pkgdown URL in DESCRIPTION's URL field and in _pkgdown.yml.  bbr
doesn't do the first yet.

Add manual links for all bbr references so to override the dead
rdrr.io links that pkgdown would otherwise fall back to.
(Unfortunately the rdrr.io are still used in the code blocks.)

Separate all of these targets at the bottom of the page so that it's
easier to reuse them and to prune them once bbr is adjusted to support
auto-linking.

[*]: https://pkgdown.r-lib.org/articles/linking.html