stanargs: don't write bbr_stanargs class to disk
get_stanargs() wraps the list in a custom child class so that it can
define a tailored print() method.  That's just for rendering purposes,
and the -stanargs.R file should contain a plain list.  And in most
cases, bbr_stanargs class does not leak out to disk because 1) the
arguments are sorted if there's more than one, and that drops the
class and 2) if .clear=TRUE, the get_stanargs() result isn't used as
the starting point.

Call unclass() on the get_stanargs() result to fix the remaining case:
when .clear=FALSE and the length of the arguments is 1.