tests: skip fxa-related tests that require latest jsonlite
cmdstanr::write_stan_json() tells jsonlite::write_json() to print with
max precision.  As of jsonlite v1.8.5, 17 significant digits are used
to ensure that the binary value is preserved when converting to base
10 and then back again.

This change in jsonlite changes the input data (in most cases),
leading to a change in the sampling results, which check_up_to_date()
rightly flags.  Once jsonlite v1.8.5 makes it to MPN, the plan is to
specify that as the minimum version for bbr.bayes, to reduces cases
where a difference in the jsonlite version is the reason for different
results.

For now, get the test suite passing with the jsonlite 1.8.5 by
skipping fxa-related tests that fail due to the changed data.

Re: #88