nmbayes: rework control stream handling
For nmbayes, the initial estimates are generated from the top-level
model.  Then the control steam is copied over to submodels (one for
each chain) and modified to uncomment lines relevant for the Bayesian
estimation run.

This method involving "live" commented code is likely to confuse
users.  Instead follow the approach suggested by @kylebaron in gh-37.
Switch to a top-level control stream that includes uncommented code
for each part.  Create a submodel for the inits like we do for the
chain submodels.  For each submodel, derive a control stream for each
submodel by dropping and adjusting the relevant parts of the top-level
control stream.

Note that the restructuring interferes with some bbi_nmbayes_model
methods falling back to the bbi_nonmem_model method.  This commit
adjusts the ones that are flagged by the tests, but others might still
require tweaks to get the desired behavior.