nmbayes: replace nmbayes_convert with copy_model_as_nmbayes
The idea behind nmbayes_convert() is that the user would typically
call copy_model_from() and then follow up with a call to
nmbayes_convert().  With stan-gq models, we took a different approach
and had copy_model_as_stan_gq() handle the copy too.  Giving the very
similar purpose of these helpers, their different approaches are
confusing.

Replace nmbayes_convert() with a copy_model_as_nmbayes() function.
Note that we're dropping much of the special handling that no longer
applies after the control stream rework in the previous commit.

Note that copy_model_as_nmbayes() will likely be extended to adjust
the control stream in some way (at least adding a comment), but that's
left for a future commit.