nmbayes: drop extrasend handling
Code in run_chains() adjusts the level of the extrasend directive.
Per Tim W, extrasend is important to support because it allows working
around NONMEM bugs (not actively used at the moment, but it has been
in the past and may be in the future).

However, the plan is to drop dedicated support and leave it to the
caller to add the extra directory level in the top-level control
stream.