copy_model_as_nmbayes: inject comment pointing to nmbayes docs
A comment is probably a good compromise between not helping new users
out at all and being too invasive.  We may want to do more, like
insert a template of the required records, but wait for scientist
feedback before going that direction.

If some users think this comment is useful but others find it
annoying, we can add an arg/option to disable it, but punt on that for
now.