copy_model_as_nmbayes: retain comments after estimation records
copy_model_as_nmbayes() deletes existing estimation records by
removing them from the nmrec_record's values.  nmrec considers
trailing comment lines to be a part of the record, so that approach
deletes those comments.

That's likely to confuse users, especially if it's a commented record,
such as

  $EST ....
  ;$TABLE ...

Add custom logic to keep the trailing comment lines.

(We may want to add a helper to nmrec, but it's not clear to me how
common this case is or what the helper should look like.)

Thanks to @seth127 for spotting this problem.