DESCRIPTION: require jsonlite 1.8.5 or later
build_data() calls cmdstanr::write_stan_json(), which requests
jsonlite write numbers with max precision.  jsonlite 1.8.5 changed the
number of significant digits that it prints in this case.

That version of jsonlite is now on the latest MPN, so declare it as
the minimum version to reduce cases where a difference in the jsonlite
version is the reason for different results.

This is a follow-up to 4ea3f82 (tests: skip fxa-related tests that
require latest jsonlite, 2023-06-09).

Closes #88.