drone: sync configuration with main
The reworked nmbayes handling has landed in main (3d68e0c).  We're not
ready to release that yet, but there are some upcoming Stan-related
changes we want to release in the meantime.  The plan is to make these
changes off of the commit from before the nmbayes rework was merged so
that the changes can be merged into a dedicated release branch (in
addition to main).

However, the PR drone build will fail because the merged tree requires
nmrec, but the drone configuration from main's side will not be used.

Pull in main's drone changes so that the PR drone build passes.

(cherry picked from commit da50a0f1be23bcd4ebe7ac38cbe5c27f1e2e3af2)