nmbayes submit_model: write table for run_sims
The upcoming run_sims() function will join a table of results to the
data.  We could rely on the user to include a table with the expected
columns at the expected spot, but, since we're modifying the control
stream for chain models anyway, add a table record for writing to a
dedicated file.