nmbayes: export chain_paths helper
get_chain_files() is unexported.  Users are likely going to want
something like it to access output files from the chain submodels.

Export a wrapper with a reduced parameter list.  The exported
functionality is a pure path computation that doesn't require the path
to be a regular file or to exist [*], so use "paths" instead of
"files" for consistency with fs and bbr::build_path_from_model().

Note that conceptually this is similar to the
bbr::build_path_from_model() generic.  Perhaps this functionality
should even be exposed as an nmbayes method for that.

However, 1) doing so has the downside of changing the return value to
a non-scalar and 2) at least currently, build_path_from_model()
doesn't have support for specifying a name to use in place of the
model ID (which chain_paths() will gain in the next commit).  So, for
now at least, expose a tailored function.

[*] though the leading chain directories must exist