run_sims: add optional progress bar for epred/ipred
In the context of simulating epred/ipred, I think ticking the progress
bar every iteration should be okay performance-wise.  If there turns
out to be a noticeable penalty, we can adjust this to only tick every
Nth iteration.  (Alternatively, with the switch to future in the next
commit, the progress bar could be moved to the batch level, but that's
probably too coarse.)