run_sims: simplify progress bar message
The progress bar is called with a message that includes the draw
number.  A few handlers display that message.  For example the
"progress" handler includes it at the right:

    - [================>-----------------]  40% EPRED (345)

It's probably confusing to show the draw numbers because theey will
have random gaps.  (The draw number is the unique index of the sample
concatenated across chains, and run_sims is randomly selecting a
subset.)

We could display n, the row of a given batch, but that's confusing
because it will reset with a new batch, and it's also not clear what
order those will come in from workers.

So just display "EPRED" or "IPRED" without any count.