run_sims: work around innocuous autonpde warnings
autonpde() allows specifying iid, ix, and iy with either a column name
or index.  Unfortunately using the name leads to three "NAs introduced
by coercion" warnings because npde calls as.integer() on each value to
decide if it is an index or name.

It'd be safe to suppress these particular warning messages, but the
downsides of doing that would be that the message 1) is coming from
base R, so there is some risk we'd silence an unrelated warning, and
2) is marked for translation.

Instead avoid the warnings by passing an index rather than name for
each of these columns.

I've verified that the `run_sims(..., ewres_npde = TRUE)` call in
test-run-sims.R returns identical results before and after this
change.

Thanks to @timwaterhouse for pinpointing the cause of these warnings.