drone: temporarily install dev mrgsolve
The nm_join_bayes tests are guarded on the mrgsolve version because
the results change with 07793922 (all records get a different draw
from EPS, 2023-08-22), which will be included in the 1.2.0 release.

Install from the mrgsolve repo that the tests aren't skipped.

Note that this will fail once the release/1.2.0 branch is deleted and
will need to be update.